refactor: Extract chart alignment helpers to another file
