+++ app/components/chart-preview.tsx
@@ -17,6 +17,10 @@ import { LoadingStateProvider } from "@/charts/shared/chart-loading-state";
 import { ChartErrorBoundary } from "@/components/chart-error-boundary";
 import { ChartFootnotes } from "@/components/chart-footnotes";
 import {
+  useAlignChartElements,
+  useChartHeaderMarginBottom,
+} from "@/components/chart-helpers";
+import {
   ChartPanelLayoutTall,
   ChartPanelLayoutVertical,
   ChartWrapper,
@@ -53,88 +57,6 @@ import { useTransitionStore } from "@/stores/transition";
 import { useTheme } from "@/themes";
 import { snapCornerToCursor } from "@/utils/dnd";
 import useEvent from "@/utils/use-event";
-
-type AlignChartElementsContext = {
-  reset: () => void;
-  maxHeaderHeight: number;
-  setMaxHeaderHeight: (height: number) => void;
-  maxChartHeight: number;
-  setMaxChartHeight: (height: number) => void;
-};
-
-const AlignChartElementsContext =
-  React.createContext<AlignChartElementsContext>({
-    reset: () => {},
-    maxHeaderHeight: 0,
-    setMaxHeaderHeight: () => {},
-    maxChartHeight: 0,
-    setMaxChartHeight: () => {},
-  });
-
-export const useAlignChartElements = () => {
-  const ctx = React.useContext(AlignChartElementsContext);
-
-  if (ctx === undefined) {
-    throw Error(
-      "You need to wrap your component in <AlignChartElementsProvider /> to useAlignChartElements()"
-    );
-  }
-
-  return ctx;
-};
-
-export const AlignChartElementsProvider = ({
-  children,
-}: {
-  children: React.ReactNode;
-}) => {
-  const [maxHeaderHeight, setMaxHeaderHeight] = React.useState(0);
-  const [maxChartHeight, setMaxChartHeight] = React.useState(0);
-  const reset = React.useCallback(() => {
-    setMaxHeaderHeight(0);
-    setMaxChartHeight(0);
-  }, []);
-
-  return (
-    <AlignChartElementsContext.Provider
-      value={{
-        reset,
-        maxHeaderHeight,
-        setMaxHeaderHeight,
-        maxChartHeight,
-        setMaxChartHeight,
-      }}
-    >
-      {children}
-    </AlignChartElementsContext.Provider>
-  );
-};
-
-export const useChartHeaderMarginBottom = () => {
-  const alignChartElements = useAlignChartElements();
-  const headerRef = React.useRef<HTMLDivElement>(null);
-
-  React.useEffect(() => {
-    if (headerRef.current) {
-      const { height } = headerRef.current.getBoundingClientRect();
-
-      if (height > alignChartElements.maxHeaderHeight) {
-        alignChartElements.setMaxHeaderHeight(height);
-      }
-    }
-  }, [alignChartElements]);
-
-  const headerMarginBottom = React.useMemo(() => {
-    if (headerRef.current) {
-      const { height } = headerRef.current.getBoundingClientRect();
-      return alignChartElements.maxHeaderHeight - height;
-    }
-
-    return 0;
-  }, [alignChartElements.maxHeaderHeight]);
-
-  return { headerRef, headerMarginBottom };
-};
 
 type ChartPreviewProps = {
   dataSource: DataSource;
